requirements.txt
crewai==1.8.0
crewai-tools==1.8.0
qdrant-client==1.16.2
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
openai>=1.3.0
langchain>=0.1.7
langchain-community>=0.0.8
langchain-openai>=0.0.5
voyageai>=0.1.0
streamlit>=1.28.0
